+++ scripts/build_and_run.sh
@@ -8,8 +8,9 @@ test_resources_dirpath="${root_dirpath}/src/test/resources/examples"
 
 extra_args="${1+"${@}"}"
 "${root_dirpath}/gradlew" build run --args=" \
-    --gov-constants ${test_resources_dirpath}/gov-constants-example.yml \
-    --scenarios ${test_resources_dirpath}/scenarios-example.yml \
-    --assets ${test_resources_dirpath}/assets-example.yml \
-    --projections ${test_resources_dirpath}/projections-example.yml \
+    --gov-constants ${test_resources_dirpath}/gov-constants.yml \
+    --scenarios ${test_resources_dirpath}/scenarios.yml \
+    --assets ${test_resources_dirpath}/assets.yml \
+    --assets-history ${test_resources_dirpath}/assets-history.yml \
+    --projections ${test_resources_dirpath}/projections.yml \
     ${extra_args}"