+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/CustomTagDefinition.java
@@ -1,0 +1,18 @@
+package com.strangegrotto.wealthdraft.assets.definition;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
+import org.immutables.value.Value;
+
+import java.util.Set;
+
+@WealthdraftImmutableStyle
+@Value.Immutable
+@JsonDeserialize(as = ImmCustomTagDefinition.class)
+public interface CustomTagDefinition {
+    // TODO Implement default values
+
+    // An empty set means all values are allowed
+    @Value.Parameter
+    Set<String> getAllowedValues();
+}