+++ src/main/java/com/strangegrotto/wealthdraft/networth/assets/AssetTag.java
@@ -1,5 +1,0 @@
-package com.strangegrotto.wealthdraft.networth.assets;
-
-public interface AssetTag {
-    String getName();
-}