+++ src/main/java/com/strangegrotto/wealthdraft/networth/projections/ProjectionScenario.java
@@ -1,6 +1,7 @@
 package com.strangegrotto.wealthdraft.networth.projections;
 
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
+import com.strangegrotto.wealthdraft.assets.temporal.AssetChange;
 import org.immutables.value.Value;
 
 import java.time.LocalDate;