+++ src/test/java/com/strangegrotto/wealthdraft/TestResourceDirnames.java
@@ -2,7 +2,8 @@ package com.strangegrotto.wealthdraft;
 
 public enum TestResourceDirnames {
     EXAMPLES("examples"),
-    ASSET_DESERIALIZATION_TESTS("assets-deserialization"),
+    ASSET_DEFINITIONS_DESERIALIZATION_TESTS("assets-deserialization"),
+    ASSETS_HISTORY_DESERIALIZATION_TESTS("assets-history-deserialization"),
     PROJECTIONS_DESERIALIZATION_TESTS("projections-deserialization");
 
     private final String dirname;